Platform fee + school platform-bills proxy + section Long fix
Backend changes:
  - PlatformFeeService fetches per-enrollment platform rate from
    platform_admin, cached 1h via Caffeine.
  - InvoiceService auto-adds a "Platform Fee (Edunix)" line on every
    student invoice crediting Platform Payable; reverts per-payment fee
    posting in cash + online flows.
  - SchoolPlatformBillsService + controller proxy /api/platform-bills
    upstream using PLATFORM_METRICS_SECRET.
  - AccountingSettings no longer requires Platform Fee Expense account.
  - InvoiceRepository.search adds invoiceNumber LIKE filter (controller
    + StudentPortalPaymentService callers updated).
  - InvoiceLine.fee_category_id relaxed to nullable for system-injected
    Platform Fee lines.
  - Migration: 2026_06_25__invoice_lines_nullable_fee_category.sql
    + update.sql with the same ALTER.

Section Long PK fix:
  - SectionRepository changed from JpaRepository<Section,Integer> to
    <Section,Long> to match Section.id.
  - Removed Math.toIntExact / .intValue coercions across
    ClassRoutineServiceImpl, SectionServiceImpl, ClassTeacherService,
    StaffClassAssignmentService, EnrollmentServiceImpl. Unblocks class
    routine creation when section IDs are large.

Config:
  - New application.properties key platform.admin.url (PLATFORM_ADMIN_URL).

+++ src/main/java/com/example/academic_service/service/StaffClassAssignmentService.java
@@ -44,7 +44,7 @@ public class StaffClassAssignmentService {
         if (req.getGenderSection() != null && req.getGenderSection().getId() != null)
             a.setGenderSection(genderSectionRepository.getReferenceById(req.getGenderSection().getId()));
         if (req.getSection() != null && req.getSection().getId() != null)
-            a.setSection(sectionRepository.getReferenceById(req.getSection().getId().intValue()));
+            a.setSection(sectionRepository.getReferenceById(req.getSection().getId()));
         if (req.getStudentGroup() != null && req.getStudentGroup().getId() != null)
             a.setStudentGroup(studentGroupRepository.getReferenceById(req.getStudentGroup().getId()));
         return assignmentRepository.save(a);
